+++ src/index.ts
@@ -2,5 +2,6 @@ export { CohostCheckoutProvider, CohostCheckoutContext, useCohostCheckout } from
 export { CohostEventProvider, useCohostEvent } from './context/CohostEventContext';
 export { CohostProvider, type CohostProviderProps } from './context/CohostContext';
 export { PaymentElementProvider, usePaymentElement } from './context/PaymentElementContext';
+export { CohostStartCheckoutProvider } from './provider/CohostStartCheckoutProvider';
 export { CreditCardInformation } from './lib/tokenizers/types';
 export { useCohost } from './hooks/useCohost';