+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cohostvip/cohost-react
 
+## 0.1.10
+
+### Patch Changes
+
+- added support for trakcing code and context when starting a new cart session"
+- Updated dependencies
+  - @cohostvip/cohost-node@0.1.10
+
 ## 0.1.8
 
 ### Patch Changes